[Delete] ViewProtocol & RectangleViewClass
