--- DrawingApp/DrawingApp/Protocol/RectangleViewable.swift
@@ -1,12 +1,0 @@
-//
-//  RectangleViewable.swift
-//  DrawingApp
-//
-//  Created by 박진섭 on 2022/03/16.
-//
-
-import UIKit
-
-protocol RectangleViewable:UIView {
-    func setupFrameWithRectangle(rect:Rectangleable)
-}
